requirements.txt
prettytable==3.9.0
python-telegram-bot==20.8